refactor(app): move sprinkles and global themes to @codeimage/ui
