+++ apps/codeimage/src/components/BottomBar/BottomBar.css.ts
@@ -1,5 +1,5 @@
 import {style} from '@vanilla-extract/css';
-import {themeVars} from '../../theme/global.css';
+import {themeVars} from '@codeimage/ui';
 import {scaffoldVars} from '../Scaffold/Scaffold.css';
 
 export const wrapper = style({
