+++ apps/codeimage/src/components/Frame/Frame.css.ts
@@ -1,6 +1,6 @@
 import {createTheme, style} from '@vanilla-extract/css';
 import {backgroundColorVar} from '../../theme/variables.css';
-import {themeVars} from '../../theme/global.css';
+import {themeVars} from '@codeimage/ui';
 
 export const [frameHandler, frameHandlerVars] = createTheme({
   scale: '1',
