+++ apps/codeimage/src/components/LoadingOverlay/LoadingOverlay.css.ts
@@ -1,5 +1,5 @@
 import {style} from '@vanilla-extract/css';
-import {themeVars} from '../../theme/global.css';
+import {themeVars} from '@codeimage/ui';
 
 export const overlay = style({
   position: 'absolute',
