+++ apps/codeimage/src/components/Scaffold/Sidebar/Sidebar.css.ts
@@ -1,5 +1,5 @@
 import {createTheme, style} from '@vanilla-extract/css';
-import {themeVars} from '../../../theme/global.css';
+import {themeVars} from '@codeimage/ui';
 import {scaffoldVars} from '../Scaffold.css';
 import {recipe, RecipeVariants} from '@vanilla-extract/recipes';
