+++ apps/codeimage/src/components/ShadowField/ShadowField.tsx
@@ -5,7 +5,7 @@ import {sprinkles} from '../../theme/sprinkles.css';
 import {DropdownMenu} from '../../ui/Dropdown/DropdownPanel';
 import {DropdownPortal} from '../../ui/Dropdown/DropdownPortal';
 import {ShadowForm} from './ShadowForm';
-import {themeVars} from '../../theme/global.css';
+import {themeVars} from '@codeimage/ui';
 
 import * as fieldStyles from '../../ui/TextField/TextField.css';
 import clsx from 'clsx';
