+++ apps/codeimage/src/components/Terminal/windows/WindowsTerminal.css.ts
@@ -1,5 +1,5 @@
 import {createTheme, style} from '@vanilla-extract/css';
-import {themeVars} from '../../../theme/global.css';
+import {themeVars} from '@codeimage/ui';
 import {header} from '../terminal.css';
 
 export const [theme, vars] = createTheme({
