+++ apps/codeimage/src/components/ThemeSwitcher/ThemeSwitcher.css.ts
@@ -1,5 +1,5 @@
 import {createTheme, createVar, style} from '@vanilla-extract/css';
-import {themeVars} from '../../theme/global.css';
+import {themeVars} from '@codeimage/ui';
 import {backgroundColorVar} from '../../theme/variables.css';
 import {recipe, RecipeVariants} from '@vanilla-extract/recipes';
