+++ apps/codeimage/src/ui/ColorPicker/ColorPicker.css.ts
@@ -1,5 +1,5 @@
 import {style} from '@vanilla-extract/css';
-import {themeVars} from '../../theme/global.css';
+import {themeVars} from '@codeimage/ui';
 
 import * as fieldStyles from '../TextField/TextField.css';
