+++ apps/codeimage/src/ui/Dialog/Dialog.css.ts
@@ -1,5 +1,5 @@
 import {createTheme, style} from '@vanilla-extract/css';
-import {themeVars} from '../../theme/global.css';
+import {themeVars} from '@codeimage/ui';
 import {fontSize, fontWeight} from '../Text/Text.css';
 import {recipe, RecipeVariants} from '@vanilla-extract/recipes';
 import {adaptiveFullScreenHeight} from '../../theme/base.css';
