+++ apps/codeimage/src/ui/Label/FieldLabel.css.ts
@@ -1,5 +1,5 @@
 import {style} from '@vanilla-extract/css';
-import {themeVars} from '../../theme/global.css';
+import {themeVars} from '@codeimage/ui';
 
 export const label = style({
   color: themeVars.dynamicColors.input.labelTextColor,
