+++ apps/codeimage/src/ui/PortalHost/PortalHost.css.ts
@@ -1,5 +1,5 @@
 import {style} from '@vanilla-extract/css';
-import {themeVars} from '../../theme/global.css';
+import {themeVars} from '@codeimage/ui';
 
 export const host = style({
   width: '0px',
