+++ apps/codeimage/src/ui/RadioBlock/RadioBlock.css.ts
@@ -1,7 +1,7 @@
 import {style} from '@vanilla-extract/css';
 import {recipe, RecipeVariants} from '@vanilla-extract/recipes';
 import * as textFieldTheme from '../TextField/TextField.css';
-import {themeVars} from '../../theme/global.css';
+import {themeVars} from '@codeimage/ui';
 import * as groupStyles from '../Group/Group.css';
 
 export const radioBlock = recipe({
