+++ apps/codeimage/src/ui/SegmentedField/SegmentedField.css.ts
@@ -1,5 +1,5 @@
 import {createTheme, style} from '@vanilla-extract/css';
-import {themeVars} from '../../theme/global.css';
+import {themeVars} from '@codeimage/ui';
 import * as textFieldStyles from '../TextField/TextField.css';
 
 export const [segmentedFieldTheme, segmentedFieldVars] = createTheme({
