+++ packages/ui/src/index.ts
@@ -1,2 +1,2 @@
-export * from './lib/theme';
-export * from './lib/primitives';
+export {colors, spacing, themeColors, themeVars, root} from './lib/theme';
+export {Button} from './lib/primitives';
