+++ packages/ui/src/lib/primitives/index.ts
@@ -1,1 +1,1 @@
-export {};
+export {Button} from './Button';
