+++ packages/ui/src/lib/theme/index.ts
@@ -1,2 +1,4 @@
 export * from './colors';
 export * from './spacing';
+export * from './theme.css';
+export * from './global.css';
