+++ packages/ui/src/lib/theme/variables.css.ts
@@ -1,0 +1,4 @@
+import {createVar} from '@vanilla-extract/css';
+
+export const colorVar = createVar();
+export const backgroundColorVar = createVar();
